arregle el tipo de durabilidad a double para que se pueda manejar con…

+++ src/fiuba/algo3/tp2/estrategiasDesgaste/EstrategiaDesgasteAbrupta.java
@@ -4,7 +4,7 @@ public class EstrategiaDesgasteAbrupta extends EstrategiaDesgaste{
 
     private int usos;
 
-    public EstrategiaDesgasteAbrupta(int durabilidad, int usos){
+    public EstrategiaDesgasteAbrupta(double durabilidad, int usos){
 
         this.durabilidad = durabilidad;
         this.usos = usos;
